% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/econdatasets-package.R
\docType{package}
\name{econdatasets-package}
\alias{econdatasets}
\alias{econdatasets-package}
\title{econdatasets: Easily Download 'EconDataverse' Datasets}
\description{
The 'EconDataverse' is a universe of open-source packages to work seamlessly with economic data. This package is designed to make it easy to download selected datasets that are preprocessed by 'EconDataverse' packages in a single step. Learn more about the 'EconDataverse' at \url{https://www.econdataverse.org}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://tidy-intelligence.github.io/r-econdatasets/}
}

}
\author{
\strong{Maintainer}: Christoph Scheuch \email{christoph@tidy-intelligence.com} (\href{https://orcid.org/0009-0004-0423-6819}{ORCID}) [copyright holder]

}
\keyword{internal}
